+++ lib/Braintree/Customer.php
@@ -14,7 +14,7 @@ namespace Braintree;
  *
  * @property-read \Braintree\Address[] $addresses
  * @property-read \Braintree\AndroidPayCard[] $androidPayCards
- * @property-read \Braintree\AmexExpressCheckoutCard[] $amexExpressCheckoutCards
+ * @property-read \Braintree\AmexExpressCheckoutCard[] $amexExpressCheckoutCards DEPRECATED
  * @property-read \Braintree\ApplePayCard[] $applePayCards
  * @property-read string $company
  * @property-read \DateTime $createdAt
@@ -26,7 +26,7 @@ namespace Braintree;
  * @property-read string $graphQLId
  * @property-read string $id
  * @property-read string $lastName
- * @property-read \Braintree\MasterpassCard[] $masterpassCards
+ * @property-read \Braintree\MasterpassCard[] $masterpassCards DEPRECATED
  * @property-read \Braintree\PaymentMethod[] $paymentMethods
  * @property-read \Braintree\PayPalAccount[] $paypalAccounts
  * @property-read string $phone
@@ -227,6 +227,7 @@ class Customer extends Base
         }
         $this->_set('applePayCards', $applePayCardArray);
 
+        // NEXT_MAJOR_VERSION rename Android Pay to Google Pay
         $androidPayCardArray = [];
         if (isset($customerAttribs['androidPayCards'])) {
             foreach ($customerAttribs['androidPayCards'] AS $androidPayCard) {
@@ -235,6 +236,7 @@ class Customer extends Base
         }
         $this->_set('androidPayCards', $androidPayCardArray);
 
+        // NEXT_MAJOR_VERSION remove deprecated AmexExpressCheckout
         $amexExpressCheckoutCardArray = [];
         if (isset($customerAttribs['amexExpressCheckoutCards'])) {
             foreach ($customerAttribs['amexExpressCheckoutCards'] AS $amexExpressCheckoutCard) {
@@ -259,6 +261,7 @@ class Customer extends Base
         }
         $this->_set('visaCheckoutCards', $visaCheckoutCardArray);
 
+        // NEXT_MAJOR_VERSION remove deprecated Masterpass
         $masterpassCardArray = [];
         if (isset($customerAttribs['masterpassCards'])) {
             foreach ($customerAttribs['masterpassCards'] AS $masterpassCard) {