+++ lib/Braintree/Transaction.php
@@ -144,7 +144,7 @@ namespace Braintree;
  * @property-read \Braintree\AddOn[] $addons
  * @property-read string $additionalProcessorResponse raw response from processor
  * @property-read string $amount transaction amount
- * @property-read \Braintree\Transaction\AmexExpressCheckoutCardDetails $amexExpressCheckoutCardDetails transaction Amex Express Checkout card info
+ * @property-read \Braintree\Transaction\AmexExpressCheckoutCardDetails $amexExpressCheckoutCardDetails DEPRECATED transaction Amex Express Checkout card info.
  * @property-read \Braintree\Transaction\AndroidPayCardDetails $androidPayCardDetails transaction Android Pay card info
  * @property-read \Braintree\Transaction\ApplePayCardDetails $applePayCardDetails transaction Apple Pay card info
  * @property-read \Braintree\AuthorizationAdjustment[] $authorizationAdjustments populated when a transaction has authorization adjustments created when submitted for settlement
@@ -172,7 +172,7 @@ namespace Braintree;
  * @property-read string $graphQLId transaction graphQLId
  * @property-read string $id transaction id
  * @property-read \Braintree\TransactionLineItem[] $lineItems
- * @property-read \Braintree\Transaction\MasterpassCardDetails $masterpassCardDetails transaction Masterpass card info
+ * @property-read \Braintree\Transaction\MasterpassCardDetails $masterpassCardDetails DEPRECATED transaction Masterpass card info
  * @property-read string $merchantAccountId
  * @property-read string $networkTransactionId
  * @property-read string $orderId
@@ -297,6 +297,7 @@ class Transaction extends Base
             );
         }
 
+        // NEXT_MAJOR_VERSION rename Android Pay to Google Pay
         if (isset($transactionAttribs['androidPayCard'])) {
             $this->_set('androidPayCardDetails',
                 new Transaction\AndroidPayCardDetails(
@@ -305,6 +306,7 @@ class Transaction extends Base
             );
         }
 
+        // NEXT_MAJOR_VERSION remove deprecated masterpassCard
         if (isset($transactionAttribs['masterpassCard'])) {
             $this->_set('masterpassCardDetails',
                 new Transaction\MasterpassCardDetails(
@@ -329,6 +331,7 @@ class Transaction extends Base
             );
         }
 
+        // NEXT_MAJOR_VERSION remove deprecated amexExpressCheckoutCard
         if (isset($transactionAttribs['amexExpressCheckoutCard'])) {
             $this->_set('amexExpressCheckoutCardDetails',
                 new Transaction\AmexExpressCheckoutCardDetails(