+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.3.1
+- Deprecate `masterpassCard` and `amexExpressCheckoutCard` payment methods
+- Deprecate `amexExpressCheckoutCardDetails`
+
 ## 5.3.0
 * Add `RISK_THRESHOLD` to `GatewayRejectionReason` constants
 * Add `networkTransactionId` to `CreditCardVerification`